% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Islam2011}
\alias{Islam2011}
\title{scRNA-seq data from Islam et al. 2011}
\format{A matrix with 22936 rows (genes) and 96 columns (samples). The first 48 columns are ES cells, the following 44 columns are mouse embryonic fibroblasts and the remaining 4 columns and negative controls. Data is in raw counts format.}
\usage{
data(Islam2011)
}
\description{
GEO accession GSE29087:
92 single cells (48 mouse embryonic stem cells, 44 mouse embryonic fibroblasts and 4 negative controls) were analyzed by single-cell tagged reverse transcription (STRT).
}
\references{
Islam et al. (2011) Genome Res 2011 Jul;21(7):1160-7. PMID: 21543516
}
