% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linnorm.DataImput.R
\name{Linnorm.DataImput}
\alias{Linnorm.DataImput}
\title{Linnorm Data Imputation Function.}
\usage{
Linnorm.DataImput(datamatrix, DIMZP = 0.5, method = "euclidean")
}
\arguments{
\item{datamatrix}{The matrix or data frame that contains your dataset. It is only compatible with log transformed datasets.}

\item{DIMZP}{Double >=0, <= 1. Genes not satisfying this threshold will be removed. For exmaple, if set to 0.3, genes without at least 30 percent of the samples being non-zero will be removed. Defaults to 0.5.}

\item{method}{Character. Method for calculating the distance matrix. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson", "correlation", "spearman" or "kendall". Any unambiguous substring can be given. Defaults to "euclidean".}
}
\value{
This function returns a data matrix.
}
\description{
This function performs data imputation for (sc)RNA-seq expression data or large scale count data. It will treat every zero count in the dataset as missing data and replace them with predicted values.
}
\details{
This function performs data imputation on the dataset. It first separates the genes into numPortion portions. Then, for each portion, it performs linear regression between the non zero values of each sample and its corresponding weighted mean experssion to predict missing values. The linear regerssion lines are continuous throughout all genes. If the number of genes of each portion is less than 100, the missing values will simply be replaced by weighted means.
}
\examples{
#Obtain example matrix:
data(Islam2011)
#Transformation:
transformedExp <- Linnorm.DataImput(Islam2011)
}
\keyword{Linnorm}
\keyword{RNA-seq}
\keyword{cell}
\keyword{data}
\keyword{imputation}
\keyword{missing}
\keyword{single}
\keyword{value}

