% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linnorm.R
\name{Linnorm}
\alias{Linnorm}
\title{Linnorm Function}
\usage{
Linnorm(datamatrix, showinfo = FALSE, method = "default",
  perturbation = 10, minZeroPortion = 2/3)
}
\arguments{
\item{datamatrix}{The matrix or data frame that contains your dataset. Each row is a feature (ex. Gene) and each column is a sample. Undefined values such as NA are not supported.}

\item{method}{"default" or "lambda" The program will output the transformed matrix if the method is "default". Otherwise, if the method is "lambda", the program will output a lambda value.}

\item{perturbation}{Integer >= 2. This is the perturbation during the Iterated Local Search algorithm, when we are searching for the global minimum of the "deviation parameter" measure for lambda (Please refer to the article). The range of the areas searched locally will be the exponents of this perturbation, from closest to furthest from the local minima found by using the whole range. Increasing this value will make the program much faster, but less capable of finding the global minimum. Defaults to 10.}

\item{minZeroPortion}{double >=0, <= 1 Rows without at least this portion of non-zero values will not be used in the calculation of normalizing parameter. Defaults to 2/3.}
}
\description{
This function performs the Linear model and normality based normalization method (Linnorm).
}
\examples{
normalizedExp <- Linnorm(expMatrix, showinfo = "TRUE")
}
\keyword{CPM}
\keyword{Count}
\keyword{Expression}
\keyword{FPKM}
\keyword{Linnorm}
\keyword{Parametric}
\keyword{RNA-seq}
\keyword{RPKM}
\keyword{Raw}
\keyword{TPM}
\keyword{normalization}
\keyword{transformation}

