% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linnorm.R
\name{NBSim}
\alias{NBSim}
\title{This function simulates a dataset with negative binomial distribution based on the input dataset.}
\usage{
NBSim(thisdata, NumRep = 3, NumDiff = 5000, NumFea = 20000,
  showinfo = FALSE, MaxLibSizelog2FC = 0.5)
}
\arguments{
\item{thisdata}{Matrix:    The matrix or data frame that contains your dataset. Each row is a feature (ex. Gene) and each column is a replicate. Undefined values such as NA are not supported. It is assumed that all samples are replicates of the same sample.}

\item{NumRep}{Integer: The number of replicates. This is half of the number of output samples. Defaults to 3.}

\item{NumDiff}{Integer: The number of Differentially Changed Features. Defaults to 5000.}

\item{NumFea}{Integer: The number of Total Features. Defaults to 20000.}

\item{showinfo}{Logical: should we show data information on the console? Defaults to FALSE.}

\item{MaxLibSizelog2FC}{Double: The maximum library size difference from the mean that is allowed, in terms of log 2 fold change. Set to 0 to prevent program from generating library size differences. Defaults to 0.5.}
}
\description{
This function simulates a dataset with negative binomial distribution based on the input dataset.
}
\examples{
simulateddata <- NBSim(expMatrix)
}
\keyword{Binomial}
\keyword{Count}
\keyword{Expression}
\keyword{Negative}
\keyword{RNA-seq}
\keyword{Raw}
\keyword{Simulation}
\keyword{distribution}

