% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linnorm.R
\name{RnaXSim}
\alias{RnaXSim}
\title{This function simulates a RNA-seq dataset based on a given distribution.}
\usage{
RnaXSim(thisdata, distribution = "Poisson", NumRep = 3, NumDiff = 5000,
  NumFea = 20000, showinfo = FALSE, MaxLibSizelog2FC = 0.5)
}
\arguments{
\item{thisdata}{Matrix:    The matrix or data frame that contains your dataset. Each row is a gene and each column is a replicate. Undefined values such as NA are not supported. This program assumes that all columns are replicates of the same sample.}

\item{distribution}{Character: Defaults to "Poisson". This parameter controls the output distribution of the simulated RNA-seq dataset. It can be one of "Gamma" (Gamma distribution), "Poisson" (Poisson distribution), "LogNorm" (Log Normal distribution) or "NB" (Negative Binomial distribution).}

\item{NumRep}{Integer: The number of replicates. This is half of the number of output samples. Defaults to 3.}

\item{NumDiff}{Integer: The number of Differentially Changed Features. Defaults to 5000.}

\item{NumFea}{Integer: The number of Total Features. Defaults to 20000.}

\item{showinfo}{Logical: should we show data information on the console? Defaults to FALSE.}

\item{MaxLibSizelog2FC}{Double: The maximum library size difference from the mean that is allowed, in terms of log 2 fold change. Set to 0 to prevent program from generating library size differences. Defaults to 0.5.}
}
\value{
This function returns a list that contains a matrix of count data in integer raw count and a vector that shows which genes are differentially expressed. In the matrix, each row is a gene and each column is a replicate. The first NumRep (see parameter) of the columns belong to sample 1, and the last NumRep (see parameter) of the columns belong to sample 2. There will be NumFea (see parameter) number of rows.
}
\description{
This function simulates a RNA-seq dataset based on a given distribution.
}
\examples{
#Obtain example matrix:
library(seqc)
SampleA <- ILM_aceview_gene_BGI[,grepl("A_",colnames(ILM_aceview_gene_BGI))]
rownames(SampleA) <- ILM_aceview_gene_BGI[,2]
#Extract a portion of the matrix for an example
expMatrix <- SampleA[,1:10]
simulateddata <- RnaXSim(expMatrix, distribution="NB", NumRep=3, NumDiff = 500, NumFea = 3000)
}
\keyword{"Log}
\keyword{"Negative}
\keyword{Binomial"}
\keyword{Count}
\keyword{Expression}
\keyword{Gamma}
\keyword{Normal"}
\keyword{Poisson}
\keyword{RNA-seq}
\keyword{Raw}
\keyword{Simulate}
\keyword{Simulation}
\keyword{distribution}

