% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEQC}
\alias{SEQC}
\title{Partial RNA-seq data from SEQC/MAQC-III Sample A}
\format{A matrix with 50227 rows (genes) and 10 columns (samples).}
\usage{
data(SEQC)
}
\description{
Raw Count data
}
\references{
SEQC/MAQC-III Consortium. A comprehensive assessment of RNA-seq accuracy, reproducibility and information content by the Sequencing Quality Control Consortium. Nature biotechnology 32.9 (2014): 903-914.
}
