% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linnorm.Cor.R
\name{Linnorm.Cor}
\alias{Linnorm.Cor}
\title{Linnorm-gene correlation pipeline for gene correlation study.}
\usage{
Linnorm.Cor(datamatrix, input = "Raw", method = "pearson",
  showinfo = FALSE, perturbation = 10, minZeroPortion = 0.5,
  keepAll = FALSE, sig.q = 0.2)
}
\arguments{
\item{datamatrix}{The matrix or data frame that contains your dataset. Each row is a feature (or Gene) and each column is a sample (or replicate). Raw Counts, CPM, RPKM, FPKM or TPM are supported. Undefined values such as NA are not supported. It is not compatible with log transformed datasets. If a Linnorm transfored dataset is being used, please set the "input" argument into "Linnorm".}

\item{input}{Character. "Raw" or "Linnorm". In case you have already transformed your dataset with Linnorm, set input into "Linnorm" so that you can input the Linnorm transformed dataset into the "datamatrix" argument. Defaults to "Raw".}

\item{method}{Character. "pearson", "kendall" or "spearman". Method for the calculation of correlation coefficients. Defaults to "pearson"}

\item{showinfo}{Logical. Show lambda value calculated. Defaults to FALSE.}

\item{perturbation}{Integer >=2. To search for an optimal minimal deviation parameter (please see the article), Linnorm uses the iterated local search algorithm which perturbs away from the initial local minimum. The range of the area searched in each perturbation is exponentially increased as the area get further away from the initial local minimum, which is determined by their index. This range is calculated by 10 * (perturbation ^ index).}

\item{minZeroPortion}{Double >=0, <= 1. For example, setting minZeroPortion as 0.5 will remove genes with more than half data values being zero in the calculation of normalizing parameter. Since this test is based on variance, which requires more non-zero values, it is suggested to set it to a larger value. Defaults to 0.5.}

\item{keepAll}{Boolean. After applying minZeroPortion filtering, should Linnorm keep all genes in the results? Defualts to FALSE.}

\item{sig.q}{Double >=0, <= 1. Only gene pairs with q values less than this threshold will be included in the results. Defaults to 0.2.}
}
\value{
This function will output a data frame with the following columns:
\itemize{
 \item{Gene1:}{ Name of gene 1.}
 \item{Gene2:}{ Name of gene 2.}
 \item{XPM1:}{ Gene 1 average expression level in XPM. If input is raw coutns or CPM, this column is in CPM unit. If input is RPKM, FPKM or TPM, this column is in the TPM unit.}
 \item{XPM2:}{ Gene 2 average expression level in XPM. If input is raw coutns or CPM, this column is in CPM unit. If input is RPKM, FPKM or TPM, this column is in the TPM unit.}
 \item{Cor:}{ Correlation coefficient between the two genes.}
 \item{p.value:}{ p value of the correlation coefficient.}
 \item{q.value:}{ q value of the correlation coefficient.}
}
}
\description{
This function first performs Linnorm transformation on the dataset. Then, it will calculate correlation coefficient for all pairs of genes.
}
\details{
This function performed gene correlated study in the dataset by using Linnorm transformation.
}
\examples{
data(Islam2011)
#Correlation of randomly chosen 100 genes
results <- Linnorm.Cor(Islam2011[sample(1:nrow(Islam2011),100),])
}
\keyword{CPM}
\keyword{Count}
\keyword{Expression}
\keyword{FPKM}
\keyword{Linnorm}
\keyword{Parametric}
\keyword{RNA-seq}
\keyword{RPKM}
\keyword{Raw}
\keyword{TPM}
\keyword{coefficient}
\keyword{correlation}
\keyword{kendall}
\keyword{normalization}
\keyword{pearson}
\keyword{spearman}
\keyword{transformation}

