% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linnorm.HClust.R
\name{Linnorm.HClust}
\alias{Linnorm.HClust}
\title{Linnorm-hierarchical clustering analysis.}
\usage{
Linnorm.HClust(datamatrix, showinfo = FALSE, input = "Raw",
  perturbation = 10, minZeroPortion = 0, keepAll = TRUE,
  method = "ward.D2", Group = NULL, num_Clust = 4, ClustRect = TRUE,
  RectColor = "red", fontsize = 0.5, linethickness = 0.5)
}
\arguments{
\item{datamatrix}{The matrix or data frame that contains your dataset. Each row is a feature (or Gene) and each column is a sample (or replicate). Raw Counts, CPM, RPKM, FPKM or TPM are supported. Undefined values such as NA are not supported. It is not compatible with log transformed datasets. If a Linnorm transfored dataset is being used, please set the "input" argument into "Linnorm".}

\item{showinfo}{Logical. Show information about the computing process. Defaults to FALSE.}

\item{input}{Character. "Raw" or "Linnorm". In case you have already transformed your dataset with Linnorm, set input into "Linnorm" so that you can input the Linnorm transformed dataset into the "datamatrix" argument. Defaults to "Raw".}

\item{perturbation}{Integer >=2. To search for an optimal minimal deviation parameter (please see the article), Linnorm uses the iterated local search algorithm which perturbs away from the initial local minimum. The range of the area searched in each perturbation is exponentially increased as the area get further away from the initial local minimum, which is determined by their index. This range is calculated by 10 * (perturbation ^ index).}

\item{minZeroPortion}{Double >=0, <= 1. For example, setting minZeroPortion as 0.5 will remove genes with more than half data values being zero in the calculation of normalizing parameter. Defaults to 0.}

\item{keepAll}{Logical. After applying minZeroPortion filtering, should Linnorm keep all genes in the results? Defualts to TRUE.}

\item{method}{Charcter. Method to be used in hierarchical clustering. (From hclust{stats}: the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).) Defaults to "ward.D2".}

\item{Group}{Character vector with length equals to sample size. Each character in this vector corresponds to each of the columns (samples) in the datamatrix. If this is provided, sample names will be colored according to their group. Defaults to NULL.}

\item{num_Clust}{Integer >= 0. Number of clusters in hierarchical clustering. No cluster will be highlighted if this is set to 0. Defaults to 4.}

\item{ClustRect}{Logical. If num_Clust > 0, should a rectangle be used to highlight the clusters? Defaults to TRUE.}

\item{RectColor}{Character. If ClustRect is TRUE, this controls the color of the rectangle. Defaults to "red".}

\item{fontsize}{Numeric. Font size of the texts in the figure. Defualts to 0.5.}

\item{linethickness}{Numeric. Controls the thickness of the lines in the figure. Defaults to 0.5.}
}
\value{
It returns a list with the following objects:
\itemize{
 \item{Results:}{ If num_Clust > 0, this outputs a named vector that contains the cluster assignment information of each sample. Else, this outputs a number 0.}
 \item{plot:}{ Plot of hierarchical clustering.}
 \item{Linnorm:}{ Linnorm transformed and filtered data matrix.}
}
}
\description{
This function first performs Linnorm transformation on the dataset. Then, it will perform hierarchical clustering analysis.
}
\details{
This function performs PCA clustering using Linnorm transformation.
}
\examples{
#Obtain example matrix:
data(Islam2011)
#Example:
HClust.results <- Linnorm.HClust(Islam2011, Group=c(rep("ESC",48), rep("EF",44), rep("NegCtrl",4)), num_Clust=3, fontsize=2)
}
\keyword{CPM}
\keyword{Clustering}
\keyword{Count}
\keyword{Expression}
\keyword{FPKM}
\keyword{Linnorm}
\keyword{Parametric}
\keyword{RNA-seq}
\keyword{RPKM}
\keyword{Raw}
\keyword{TPM}
\keyword{hierarchical}
\keyword{normalization}
\keyword{transformation}

