% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linnorm.R
\name{Linnorm.PCA}
\alias{Linnorm.PCA}
\title{Linnorm-PCA Clustering pipeline for Single Cell Sub-population Analysis}
\usage{
Linnorm.PCA(datamatrix, showinfo = FALSE, perturbation = 10,
  minZeroPortion = 0.5, num_center = c(1:20), Group = NA,
  pca.scale = FALSE, kmeans.iter = 2000)
}
\arguments{
\item{datamatrix}{The matrix or data frame that contains your dataset. Each row is a feature (or Gene) and each column is a sample (or replicate). Undefined values such as NA are not supported.}

\item{showinfo}{Logical. Show information about the computing process. Defaults to FALSE.}

\item{perturbation}{Integer >=2. To search for an optimal minimal deviation parameter (please see the article), Linnorm uses the iterated local search algorithm which perturbs away from the initial local minimum. The range of the area searched in each perturbation is exponentially increased as the area get further away from the initial local minimum, which is determined by their index. This range is calculated by 10 * (perturbation ^ index).}

\item{minZeroPortion}{Double >=0, <= 1. Only features (genes) with more than this portion of non-zero values will be used in the calculation of normalizing parameter. Defaults to 0.5.}

\item{num_center}{Numeric vector. Number of clusters to be tested for k-means clustering. fpc, vegan, mclust and apcluster packages are used to determine the number of clusters needed. If only one number is supplied, it will be used and this test will be skipped. Defaults to c(1:20).}

\item{Group}{Character vector with length equals to sample size. Each character in this vector corresponds to each of the columns (samples) in the datamatrix. This is for plotting purposes only. In the plot, the shape of the points that represent each sample will be indicated by their group assignment. Defaults to NA.}

\item{pca.scale}{Logical. In the prcomp(for Principal component analysis) function, set the "scale." parameter. It signals the function to scale unit variances in the variables before the analysis takes place. Defaults to FALSE.}

\item{kmeans.iter}{Numeric. Number of iterations in k-means clustering. Defaults to 2000.}
}
\value{
It returns a list with the following results:
\itemize{
 \item{k_means:}{Output of kmeans(for K-means clustering) from the stat package. Note: It contains a "cluster" object that indicates each sample's cluster assignment.}
 \item{PCA:}{Output of prcomp(for Principal component analysis) from the stat package.}
 \item{plot:}{Plot of PCA clustering.}
 \item{Linnorm:}{Linnorm transformed and filtered data matrix.}
}
}
\description{
This function first performs Linnorm transformation on the dataset. Then, it will perform Principal component analysis on the dataset and use k-means clustering to identify subpopulations of cells.
}
\details{
This function performs PCA clustering using Linnorm transformation.
}
\examples{
#Obtain example matrix:
data(Islam2011)
#Example 1
PCA.results <- Linnorm.PCA(Islam2011)
}
\keyword{Analysis}
\keyword{CPM}
\keyword{Clustering}
\keyword{Component}
\keyword{Count}
\keyword{Expression}
\keyword{FPKM}
\keyword{K-means}
\keyword{Linnorm}
\keyword{PCA}
\keyword{Parametric}
\keyword{Principal}
\keyword{RNA-seq}
\keyword{RPKM}
\keyword{Raw}
\keyword{TPM}
\keyword{k-means}
\keyword{kmeans}
\keyword{normalization}
\keyword{transformation}

